Add files via upload
AI powered Emotion Detector can contribute to the educational field by reading children's emotions as they study and improve the learning system as well. It can be used to monitor students’ engagement with their homework. It can also be used in situations with much higher stakes, such as in hiring, by airport security to flag faces as revealing deception or fear.